Add files via upload
IVP LAB - Experiment No - 4
Module - MPEG
(Posted on <26 April, 2024>, Submission Deadline <02 May, 2024>)
Problem Statement
The objective of this experiment is to evaluate the impact of MPEG
compression on image file size and visual quality. The program should
prompt the user to input parameters such as the binary or color image.
Requirements:
1. User Input:
The program should prompt user to take a 5*5 matrix as a binary image or
any color image.
2. Error Handling:
Ensure proper error handling for invalid user inputs, such as non-numeric
entries or mismatched array dimensions.
3. Output:
Display the input as well as the result of the code
Upload code & results snippets along with your comments on results and
submit it on Google Classroom page.


Rubric for Grading Submission
You need to submit your report for evaluation on a scale of 10. The report shall
consist of your understanding of the problem, code, results, novelty (if any) and
conclusions.